Issue Fix: Crash Paint Scene Reload
Issue identified: When loading a paint texture, deloading it and reloading the scene the game crashes due to a non-existant reference to a DS_Priority.

Theory: Issue might be caused due to an old reference in the event management system. Bad memory management.

Fix: When the oPaintManager is destroyed the instance that's about to be destroyed is unsubscribed from all the events it related to.